requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
async-timeout==4.0.3
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
dataclasses-json==0.6.7
docx2txt==0.9
emoji==2.15.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.3
filetype==1.2.0
frozenlist==1.8.0
gigachat==0.2.0
greenlet==3.4.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jq==1.11.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.30
langchain-gigachat==0.5.0
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.32
llvmlite==0.47.0
lxml==6.0.4
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nltk==3.9.4
numba==0.65.0
numpy==2.2.6
olefile==0.47
orjson==3.11.8
packaging==26.1
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
pypdf==6.10.2
pypdfium2==5.7.0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.26
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.5
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
six==1.17.0
soupsieve==2.8.3
sqladmin==0.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
unstructured==0.18.32
unstructured-client==0.42.12
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
webencodings==0.5.1
wrapt==2.1.2
WTForms==3.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
